+++ src/main/java/parser/EmptyInputException.java
@@ -1,4 +1,4 @@
-package BabyGronk;
+package parser;
 
 public class EmptyInputException extends Exception {
     public EmptyInputException(String message) {
