+++ src/main/java/parser/Instruction.java
@@ -1,4 +1,4 @@
-package BabyGronk;
+package parser;
 
 import java.util.Arrays;
