+++ src/main/java/parser/InvalidInputException.java
@@ -1,4 +1,4 @@
-package BabyGronk;
+package parser;
 
 public class InvalidInputException extends Exception {
     public InvalidInputException(String message) {
