+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package BabyGronk;
+package task;
 
 public abstract class Task {
     private String task;
